+++ discography_webapp/pilot_feature.py
@@ -1,3 +1,0 @@
-# TODO: [Library] Pilot Autonomous Verification
-def pilot_test():
-    pass